Separate string parser tests into their own file
